Update src/com/lavacraftserver/HarryPotterSpells/Spells/Expelliarmus.java
Cleaning up expelliarmus and improving directional throwing. It should now take the item in their hand and launch it in the direction the player is looking as well as remove the weapon from the players hand. I have not tested and this code may contain errors since I did it in notepad...